Integrate dahsboard with settings
