+++ src/pages/Settings.tsx
@@ -13,6 +13,7 @@ import {
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import { ResizableBox } from "react-resizable";
 
+import Header from "../components/Header";
 import { settingsComponentRegistry } from "../registry/settings-component-registry";
 import SettingsTypeWrapper from "../components/SettingsTypeWrapper";
 import useAppStore from "../store/useAppStore";
@@ -47,70 +48,75 @@ const Settings: React.FC = () => {
   };
 
   return (
-    <Box
-      sx={{
-        display: "flex",
-        height: "calc(100vh - 64px)",
-      }}
-    >
-      <ResizableBox
-        width={drawerWidth}
-        height={window.innerHeight - 64}
-        minConstraints={[300, Infinity]}
-        maxConstraints={[500, Infinity]}
-        resizeHandles={["e"]}
-        style={{
-          height: "100%",
-          overflow: "hidden",
-          backgroundColor: "#f5f5f5",
+    <>
+      <Box sx={{ width: "100%", zIndex: 1000 }}>
+        <Header />
+      </Box>
+      <Box
+        sx={{
+          display: "flex",
+          height: "calc(100vh - 64px)",
         }}
       >
-        {settingsKeys.map((category) => (
-          <Accordion key={category} disableGutters square>
-            <AccordionSummary expandIcon={<ExpandMoreIcon />}>
-              <Typography variant="subtitle1" fontWeight="bold">
-                {category}
-              </Typography>
-            </AccordionSummary>
-            <AccordionDetails>
-              <List disablePadding>
-                {Object.keys(settings![category]).map((subKey) => (
-                  <ListItemButton
-                    key={subKey}
-                    selected={
-                      selected?.parent === category &&
-                      selected?.child === subKey
-                    }
-                    onClick={() => handleChildClick(category, subKey)}
-                    sx={{ borderBottom: "1px solid #f8f8f1" }}
-                  >
-                    <ListItemText primary={translateToTitleCase(subKey)} />
-                  </ListItemButton>
-                ))}
-              </List>
-            </AccordionDetails>
-          </Accordion>
-        ))}
-      </ResizableBox>
+        <ResizableBox
+          width={drawerWidth}
+          height={window.innerHeight - 64}
+          minConstraints={[300, Infinity]}
+          maxConstraints={[500, Infinity]}
+          resizeHandles={["e"]}
+          style={{
+            height: "100%",
+            overflow: "hidden",
+            backgroundColor: "#f5f5f5",
+          }}
+        >
+          {settingsKeys.map((category) => (
+            <Accordion key={category} disableGutters square>
+              <AccordionSummary expandIcon={<ExpandMoreIcon />}>
+                <Typography variant="subtitle1" fontWeight="bold">
+                  {category}
+                </Typography>
+              </AccordionSummary>
+              <AccordionDetails>
+                <List disablePadding>
+                  {Object.keys(settings![category]).map((subKey) => (
+                    <ListItemButton
+                      key={subKey}
+                      selected={
+                        selected?.parent === category &&
+                        selected?.child === subKey
+                      }
+                      onClick={() => handleChildClick(category, subKey)}
+                      sx={{ borderBottom: "1px solid #f8f8f1" }}
+                    >
+                      <ListItemText primary={translateToTitleCase(subKey)} />
+                    </ListItemButton>
+                  ))}
+                </List>
+              </AccordionDetails>
+            </Accordion>
+          ))}
+        </ResizableBox>
 
-      {/* Drawer-aware content */}
-      <Box sx={{ flex: 1, p: 3 }}>
-        {settings &&
-          selected &&
-          (() => {
-            const Component =
-              settingsComponentRegistry[selected.parent]?.[selected.child];
-            if (!Component) return <Typography>{"Nothing found"}</Typography>;
-            return (
-              <SettingsTypeWrapper
-                settingsName={selected.child}
-                data={settings[selected.parent][selected.child]}
-                Component={Component}
-              />
-            );
-          })()}
+        {/* Drawer-aware content */}
+        <Box sx={{ flex: 1, p: 3 }}>
+          {settings &&
+            selected &&
+            (() => {
+              const Component =
+                settingsComponentRegistry[selected.parent]?.[selected.child];
+              if (!Component) return <Typography>{"Nothing found"}</Typography>;
+              return (
+                <SettingsTypeWrapper
+                  settingsName={selected.child}
+                  data={settings[selected.parent][selected.child]}
+                  Component={Component}
+                />
+              );
+            })()}
+        </Box>
       </Box>
-    </Box>
+    </>
   );
 };
